# enhancing stackcollapse-perf.pl "--inline" option: caching addr2line and getting rid of '??' in graphs