This PR solves two issues when collapsing stacks produced by linux perf script with `--inline` option – that is, excessive use of `addr2line` and outputting `;??;` or `;;` in the results.

#### caching addr2line results
`stackcollapse-perf.pl` calls `addr2line` on every single input line. This takes lots of time. The input lines obviously repeat most of the time. So it's a perfect place to add a cache of *module,pc -> resolved addr2line result*. This saves a ton of time:
```
$ # before patches (I lost my patience after about 20 minutes, thus ^C)
$ time perf script -i /tmp/perf_0.data | ~/bin/FlameGraph/stackcollapse-perf.pl --inline &> /dev/null
^C
real    19m29,380s
user    15m5,214s
sys     4m29,142s
$
$ # after patches
$ time perf script -i /tmp/perf_0.data | ~/bin/FlameGraph/stackcollapse-perf.pl --inline &> /dev/null
real    0m53,448s
user    0m39,918s
sys     0m15,196s
```

#### dealing with addr2line failures
Sometimes addr2line produces unhelpful results, such as:
- `??` (literally two question marks) - in my case at some intrinsics, but this can obviously happen in many other cases
- `` (empty result)  - in my case in the output I saw lines such as: 
`addr2line: /lib/modules/5.11.6-1-default/kernel/fs/btrfs/btrfs.ko.xz: file format not recognized`
and this also can be triggered by missing executable among others

This should be avoided, and I think the best is to fall back to printing default symbols.


#### .

(I guess I should have made two separate PRs, but I let it be since both fix the same `--inline` switch)