@pveentjer 2021-06-14
Thanks for making the fix. I'm suffering from the same problem. Waiting for the stackcollapse-perf takes ages due to addr2line calls.